write_version_header: don't try to resolve the tag
This change makes the version code always v9.0-abcdef,
regardless of building on a tag or on a branch. This makes
the version string more consistent in the various cases.
Removing the query of the tag avoids circular problems when
releasing the SDK where the script queries the tag by looking
at CHANGELOG but the tag doesn't exist yet.


Bug: 198293659
Change-Id: I91f40c1e14656c8282a98a6ee75162623f2fae03