Autoupdate of 2 submodules
*frameworks/base updated to 0f2729a96e0c225ca389d3b81ea172e252545c53
Merge "Parcel: markForBinder"
*system/tools/aidl updated to a4b6b292d231966d431c2055fc694f21eac3c45a
aidl: support RPC in Java

In Java, when need to make a transaction on an RPC binder object, the
Parcel needs to be in a different format. This adds a build option to
enable RPC to work (by attaching the binder to a parcel, the correct
format can be used). In the future, we could enable this by default
on all AIDL interfaces which are built against a new enough API version
(b/175819535).

Bug: 175814583
Test: atest aidl_integration_test, manual
Change-Id: Ia5234e58d0f1731ddb7bb577c4e193a81779e6ec