ANDROID: Tweak how CONFIG_ARCH_MISC_INIT is enabled
Currently this feature only works on x86, but it's a core feature and
U-Boot Kconfig doesn't protect it from being enabled on boards that
don't support it. So, enable it globally first, then disable it in the
arm.fragment, which is merged afterwards. This safely disables the
feature on our ARM targets for now.

Longer term, we should be extracting the command line from the root
FDT passed into U-Boot on ARM64, from the 'chosen' node.

Bug: 199479310
Change-Id: I68e7517ea1470236cb851f22740bec6dba192ef3