libion: Increase allocation size in large allocation test
The libion unit tests contain a test to check that system heap will fail
an unusually large allocation request. However, the allocation size used
(3GB) is insufficient to cause an allocation failure on high-end
devices. This patch queries the total RAM size of the device and issues
an allocation request for the entire size of the RAM.

Bug: 157702977
Test: atest ion-unit-tests
Change-Id: I2d2f4c57c3a7aa7baba84048b1f74019c7ffad7a
Signed-off-by: Hridya Valsaraju <hridya@google.com>