Autoupdate of 2 submodules
*build/bazel updated to 27487feb49cdb349e2a06bf35fd0c10a75b84682
Add params for non-exported includes

Previously these had to be added to copts in order to not be propagated.
However, this required a path for both the directory itself and the
directory prepended with $(BINDIR); while this functioned correctly, it
is not friendly to users writing a BUILD file.

Test: build/bazel/ci/mixed_libc.sh
Test: build/bazel/ci/bp2build.sh
Change-Id: I33145d34de75501b9f2c008e6318cd43017f91ee

*build/soong updated to 04e0a1551d87dcc553aa3d0df0c6d837cd23483e
Merge "Split local/absolute include into attributes"