Add logs on invalid test probe URLs
Log when a probe URL is present but expired, and when a probe URL has
invalid format.

This is helpful to investigate test failures where the test URL is not
being applied as expected.

The modified code will only run on devices that have
TEST_URL_EXPIRATION_TIME set, so not on production devices, unless they
were used to run automated tests (and the tests did not cleanup
properly), or the user used adb comments to set the values.

Test: atest NetworkMonitorTest CaptivePortalTest
Change-Id: Ifea4fb23e01f4f43eed6508b2ce09c858446d818