Autoupdate of 3 submodules
*external/rust/crates/libsqlite3-sys updated to 8e5c8e8167f0d53d8049ec4d60c9ce2bba8a6e00
Add liblibsqlite3_sys_noicu

The library is a variant of liblibsqlite3_sys which uses libsqlite built
with no ICU extension.

Bug: 199674764
Test: m

Change-Id: I0a80f0005f747a9ab92d1b3b7e335b24252f4a1b

*external/rust/crates/rusqlite updated to 4dd7486c10c3b203519093d1be90dbe5e2b51579
Add librusqlite_noicu

librusqlite_noicu is a variant of librusqlite which is linked to
libsqlite built with no ICU extension support. This variant is intended
to be used in environments where the ICU extension is not needed and the
dependency to ICU (which is as large as 60+ MB) is not desirable, like
microdroid.

Bug: 199674764
Test: m
Change-Id: I7a1412baa449606939dd291f51bccc871f306ec1

*external/sqlite updated to 6ca691466a2d3d3ff4343eceb1e69549044e8d4e
Merge "Add libsqlite_static_noicu"