Don't put extra bits when copying bootconfig from the disk
This change fixes a bug that extra bits are copied when the bootconfig
parameters are read from the disk. Previously, the entire blocks from
the disk were read, and they all were appended to the bootconfig area.
In practice, the size of the actual bootconfig parameters is smaller
than the disk size, because the space is allocated in block unit (512
bytes) on the disk. Copying the entire block was not only unnecessary,
but also nullified subsequent attempts to append more bootconfig
paramters because the extra bits copied (which usually would be zero)
act as null terminator.

Fixing this bug by copying only the content (a multi-line string) from
the disk.

Bug: 200106555
Test: run android.virt.test.MicrodroidTestCase#testDebugMode
Change-Id: If4869729fb6734bb9a88b9b36f7594f85911692a