Add notification button for starting new trace
This change adds a button on the "Attached trace to bug report" dialog
that can be used to quickly start a new trace. A non-exported receiver
is added to handle the broadcast that will start a trace.

Bug: 188878036
Test: Checked that bugreport-terminated traces attach the "Start new
      trace" button to the "Attached trace..." notification, and that it
      correctly starts a new trace when used. (This was done on an
      internal build, as this change only affects devices with
      BetterBug)
Change-Id: I27ea6d8dada2c6d11a3b30de2785d5018ad18dbf
Merged-In: I27ea6d8dada2c6d11a3b30de2785d5018ad18dbf