Autoupdate of 2 submodules
*build/bazel updated to d0fbbb00d1cfa8600f557d28cfe150a0e8e73be9
Remove `Linkstatic` field for static cases

With the explicit implementation of cc_library_shared,
this is now redundant for both, so cc_library_static
won't be passed it as part of kwargs. Instead, we
simply always pass it to native.cc_library as True.

Test: cc_library_static_conversion_test.go
Test: mixed_{libc,droid}.sh
Change-Id: I1189466b4cac627e2db31e83687ccca23264791f

*build/soong updated to ac5097fcf4cdc14932b21b92c7c99b125c7efac7
Create bp2build converter for cc_shared_library

Factor out shareable code with cc_static_library

Remove `Linkstatic` field for shared cases,
because it's redundant with the rule

Bug: 198241907
Test: cc_library_shared_conversion_test.go
Test: mixed_{libc,droid}.sh
Change-Id: I5be3b66f812893bce41edb425fbd143ab22d9051