Fix art compiling against musl
Fix various issues when compling art against musl:

Check for ANDROID_HOST_MUSL when defining strlcpy and sig*64.

Remove error.h, its not used and doesn't exist in musl.

Include time.h that was being included transitively in glibc but
not in musl.

Add the musl libc library name to the list.

pthread_sigqueue takes a sigval, not a sigval_t.

Bug: 190084016
Test: m USE_HOST_MUSL=true
Change-Id: I2548ef0013234c42efded411c733450c19434855