Add extra warnings.
Add the -Wextra for all targets, but also add -Wconversion for
the main code.

There was a bug where readlink was being assigned to a size_t and
this option would have detected that problem. So adding these
options to avoid this in the future.

Fix the miscompares found with the new options.

Test: Compiles. Unit tests pass.
Change-Id: I8dcae71ffc29275308e16d4178edab6dfc908206