Autoupdate of 4 submodules
*build/soong updated to 8b1add6dd471fe8699edc1cc370621172e131c4d
Merge "Add back the api_srcs argument"
*frameworks/libs/modules-utils updated to 95f0a70c711ffb18fd31ceba31aa29a430e7796e
Add a new filegroup for metalava annotations

The source of these annotations need to be present when running metalava
in the mode where it "enhanced documentation". Make a new public
filegroup for them, but retain the more restricted visibility of the
framework-annotation filegroup.

Bug: 186197911
Test: m
Change-Id: I2c09c4a8c79dc94d5070c253109b08d515e8cf19

*packages/modules/common updated to ee390f26eb899a8044a663ba7309d6de78fbb717
Enhance documentation in mainline stubs

Metalava has a feature to "enhance documentation" which is mostly based
on auto-generating documentation text based on source-retention
annotations. This is turned on automatically for doc stubs, but here
we are enabling it for module's jar stubs as well.

The point of this is to be able to replace the module source code with
module jar stub source code when generating doc stubs -- something
previously not possible because most source-retention annotations are
not present in the jar stubs.

In order for the enhancement to work properly, metalava needs to be
passed the annotation source code rather than the annotations library.
This is because of various javadoc tags controlling what documentation
is generated, e.g. @paramDoc, @returnDoc etc.

Bug: 186197911
Test: manually inspect module stub source code is enhanced
Change-Id: I3fda4cd64e49c1da33fe614c30aafd9fa0dba3b9

*tools/metalava updated to 24e7e357dea21839290e906c3850cb98b98aa3f0
Merge "Add option to enhance documentation"