Use `CLOCK_BOOTTIME` for `Instant`
Discussion is ongoing upstream at
https://github.com/rust-lang/rust/issues/87906
about how exactly to make a change like this, but most of the concerns
(backwards compatibility, legacy kernels prior to 2.6.39, non-Linux
OSes) are not applicable to us.

If upstream decideds to keep `Instant` as `CLOCK_MONOTONIC`, we should
figure out an alternate solution for our use of `quiche` before backing
this out.

Bug: 200694560
Change-Id: I188ad48bbaccd2fb0196e8273b307a7bbfc737f1