Add support for zephyr modules
Adding the zephyr/module.yml file enables us to introduce CHRE into
Zephyr as a module. The build rules (Kconfig and CMakeLists.txt) will be
owned by zephyr in order to simplify the development on CHRE. Zephyr
will point its west manifest (similar to repo's manifest) to a specific
SHA of CHRE and it will be up to the developer updating the SHA to make
sure that the Zephyr build files are still compatible with the updated
version.

Test: none needed, this doesn't affect the local repo

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I896ae5631939b46891f2d93910563ca79dbbd13c