Autoupdate of 3 submodules
*build/soong updated to b13daf2ba7135f25f81fb3b51b147c63ccda143d
Merge "Add java_binary_host support for Windows"
*prebuilts/r8 updated to 84e89a20484ad6a15f10c8294f751627f7fd450d
Enable d8 java_binary_host for Windows

So that it can be distributed automatically as part of the build-tools
sdk-repo package (which is semi-manually pulling these files today).

Bug: 187222815
Change-Id: I91cf97a6bcc0d4cd44cee5caa6fa7e3611e35ead

*tools/apksig updated to cb5e16ea45459c2cebfc532b45a0a185c124c34a
Enable apksigner for Windows

As it's already being distributed in the sdk-repo build-tools windows
package.

Bug: 187222815
Change-Id: I72cb642d11e33b198fcfa914a3b49df1b856f416