Validate @Enforce expression earlier
As suggested by Jooyung in a previous change, validate the expression
when parsing the interface. Add unit tests to confirm that the expected
errors are raised for empty or malformed expressions. With this
implementation, the condition expression is parsed twice (once when
loaded and once when generating the Java code).

Test: out/host/linux-x86/nativetest64/aidl_unittests/aidl_unittests
Bug: 197828948
Change-Id: I8d6008c33b17dc1b0e473c6299cd601f86d9629a