Fix target_cpu value in android_cc_toolchain.
This variable is used for the output directory of the SolibSymlink
action, e.g.

action 'SolibSymlink
_solib_<target cpu>/libbionic_Slibm_Sliblibm_Ubp2build_Ucc_Ulibrary_Ushared_Uunstripped.so'
  Outputs: [bazel-out/k8-fastbuild/bin/_solib_<target cpu>/libbionic_Slibm_Sliblibm_Ubp2build_Ucc_Ulibrary_Ushared_Uunstripped]

Before this change, the <target cpu> value was hardcoded, so the
outputs were clobbering each other.

This change also explicitly unsets the unused parameters to the
create_cc_toolchain_config_info constructor -- those values don't seem
to be used in the actions at all. Verified with aquery of //bionic/...

Test: CI, bp2build.sh of all 4 architectures
Bug: 200781945
Change-Id: Ic230a8f19868d7a408a10a16a029b522ca0d0055