Autoupdate of 2 submodules
*art updated to afe94ef38012142e7aaca6b23bcb4f1ca5200d31
Include ramdisk image variant of jni_headers in sdk

Bug: 195754365
Test: m art-module-sdk
Change-Id: I024bff21309f006ea894aa809c2e03b31b4bc864

*build/soong updated to 12a0a310a76b8219e7cbb5cb89edd1384c7c0b42
Add ramdisk image sdk trait to cc_library_headers

Allows an sdk to require that a cc_library_headers module provides a
ramdisk image variant for the prebuilt.

Previously, "ramdisk_available: true" would be set in the generated
prebuilt snapshot for any sdk member that specified
"ramdisk_available: true" in the source module. This change will only
add that setting to the snapshot if the ramdisk image variant trait
was explicitly requested for a member.

Bug: 195754365
Test: m nothing
Change-Id: I21e0dd02d380beabf228a72c463fcc603c77d87f