Add --min_sdk_version flag to the compiler
The flag is used to set the minimum SDK version that the generated code
should support. This will allow the compiler to generate code that
makes use of newer platform features (e.g. new APIs) when it is
guaranteed that the generated code always runs on such platforms.

For now, this flag is no-op.

Bug: 175819535
Test: aidl_unittests
Change-Id: I5297f940b0327f3e057db984b6efb454266e4923