Autoupdate of 3 submodules
*external/rust/crates/thiserror updated to 48af3adb9f409299c64b320eebba171ade47560a
Grant visibility to the CompOS APEX.

This is needed as a dependency of cxx.

Bug: 186126194
Test: Builds
Change-Id: I31c6aced2ad1a164883f04164a318aeb02621f04

*external/rust/cxx updated to 149626b05cc2cad2a79bc24dbf5cd7cab8140fb7
Grant visibility to the CompOS APEX.

Bug: 186126194
Test: Builds
Change-Id: I4a5a1697835625a526471461edbae1afb77f667a

*packages/modules/Virtualization updated to 14f0739a88e9502953b87d5e2da0231568d0e95d
Add native code for cert munging.

Compsvc returns a self-signed cert (from Keystore), but we only want
the public key. Extracting this in Rust is non-trivial, so instead we
use existing support in BoringSSL in native code. (The details are
copied from compos_key_cmd.cpp, which in turn were copied from the
now-deleted FakeCompOS in odsign.)

We could alternatively do this in compsvc itself, but I was slightly
more reluctant to introduce native code there.

Bug: 186126194
Test: Run composd_cmd twice, check it accepts the key pair it generated.
Change-Id: I3faab9a7ada149d7f2776c2fb4d2656837c95e6f