Fix testSmsScheme flake
The flake seems to be caused by the previous test,
testUnsupportedScheme, which is starting the browser instead of properly
mocking the ACTION_VIEW intent.

The previous workaround for mocking an intent sent in onDestroy did not
work, as returning null in the respondWithFunction callback causes the
intent to be sent instead of mocked.

Use a different workaround for mocking the intent sent in onDestroy.

Bug: 200766449
Test: atest CaptivePortalLoginActivityTest --rerun-until-failure 20
Change-Id: I251c38cd3023dc58e6fe1cccff317cf29c6467d9