DoH: Lock access to DohDispatcher
Previously, we exposed &mut APIs to interacting with DohDispatcher. This
is fine, but is not a legal API to use in a multithreaded context
without a lock. To support this usage, we wrap it in a Mutex and only
require a & reference which can be safely materialized from multiple
threads since Mutex is Sync.

Bug: 202080500
Change-Id: I7eb8dbaa74202033a8aa89fdedade792070aa483