verifier: Clean up `RegisterLine::SetRegisterType*()`.
In `RegisterLine::SetRegisterTypeWide()`, replace the pair
check with a DCHECK() as all callers pass valid pairs.

In `RegisterLine::SetRegisterType()`, the non-pair check was
needed only by a single caller (the only one that checked
the return value), so move the check to that caller and use
a DCHECK() instead.

Remove the now unnecessary `verifier` argument from these two
functions and change return type to `void`.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 28313047
Change-Id: Iabeb6fdb61d3bf2713bf7c5a72288fff54a35754