Monitor system configuration change
The RRO generation does not happen during package installation
and instead after the OMS receives the PACKAGE_ADDED event.
There is a delay where the old idmap is used with the new
target package resulting in a random failure for getting the
overlay in NetworkStack. Given that the fix in OMS and PM may
not be landed in short time, listen the configuration change
to re-query the resource to address this issue as a short
term solution.

Bug: 189738017
Bug: 181919957
Test: atest NetworkStackTests
Change-Id: I9a1ac0a6140fc8e9362324dc99667c7f9fa4c9d5