Do not suppress crash dialogs in the launch test
Previously crash dialogs were suppressed in the launch test and the errors were collected inside the instrumentation test through a hashmap that stores the suppressed crashes. This is not necessary because the suppressor only suppress app crash and app ANR, in which cases the system will record the errors in dropbox anyway. As a result, when a current launch test failed, we see duplicated error messages in the test log since the launch instrumentation test finds the error in both the hashmap and the dropbox. Furthermore, our users may not want to have the dialog suppressed and may want to see the dialog on the device screen. Therefore, when this CL removes the crash dialog suppression, it not only does not affect the test results, but also makes the app behaves more like in the real world in case of a crash.

Test: make csuite && csuite-tradefed run app-launch-test with a passing and failing test.
Change-Id: I9b09fef20da3204cbc5304d2ccca92e23fe85418