Loosen class status check after resolving.
If `ClassLinker::LinkClass()` creates a new `Class` object,
it takes an `ObjectLock` on that object but releases that
lock when it returns to the `ClassLinker::DefineClass()`.
Therefore another thread can try and fail to verify the
class and mark it as resolved erroneous. The `CHECK()` that
the class is not resolved erroneous was therefore wrong
and we remove that part.

Also dump the class status if the remaining `CHECK()` fails.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 202091046
Change-Id: Ida9c0d45478fbae913fcaabcc237048d28a4dd3d