Autoupdate of 2 submodules
*build/soong updated to 97b8531492fce820a61d11c2ca5ad2d5abe21eaa
Add bp2build support for cpp_std.

This converts cpp_std and gnu_extensions into a -std copt, if cpp_std is
specified or gnu_extensions is false if cpp_std is not specified.

I chose to go with this copts approach because the tradeoff is a much
simpler setting than adding a new attr(s) everywhere that uses features
to set the flag.

This approach limits the number of user-configurable knobs (since users
would then be able to set std in _both_ copts and the new attr). But it
does rely on the user copt overriding the toolchain's default gnu++17
version, which can mean a `-std` flag showing up twice in the action.

Fixes: b/202462232
Test: b build //system/libziparchive:libziparchive
Change-Id: I81dad029059461739b91f318d662e089edb46b84

*prebuilts/clang/host/linux-x86 updated to cfc620128763e4a68933b47615cf67b08f19f066
Always set the default -std copt.

Soong uses cpp_std and gnu_extensions, but Bazel simplifies it to a
regular copt on the target itself (either c++ or gnu++ or experimental).
If the copt exists on the target, it takes precedence over the compiler
one.

Test: CI
Change-Id: Ia5fa0fda636acb4e12c679d6b1b641de270e1e11