Move test into nested class
When using @RunWith(Enclosed.class), tests in the outer
class won't be run. There was even a warning comment
about this on this test class, but it was still missed.

Errorprone is being updated to 3.8.1, and adds a check for
this.

Bug: 190944875
Test: Presubmits
Change-Id: Ifd7534ed81b93182969297e7b75178fa4201ef13
Merged-In: Ifd7534ed81b93182969297e7b75178fa4201ef13
(cherry picked from commit c5863702530d254a5246f87280c534107c20c35f)