Clean up common file test usage.
A few tests had a copy of the compressed and uncompressed shared
libraries. Remove all of the uncompressed copies.

Also, some tests had maps and output that pointed to the common
directory. Replace with a special links.txt file that will create
the links at runtime.

Finally, a few tests had maps files that only had the read-execute
map for libart.so. Add the appropriate read-only map so that the
output doesn't have the bad libart.so!libart.so lines.

Test: All unit tests pass.
Change-Id: I6e0553eccf678126be2cb7696ec38ef8ab7ed504