Increase max execution time for a VM
We already increased the per-process time, now additionally increase
the overall time. This is only needed for nested virtualization, but
as before we currently do it for any VM. This is being fairly cautious
- I've only seen the timeout exceeded once, but we don't want flaky
tests.

I also modified the logic for deciding the timeout for any individual
process. I think the modification reflects the original intent, but I
may have completely misunderstood.

Bug: 200924405
Test: atest art_standalone_odrefresh_tests
Test: atest ComposTestCase
Change-Id: I2e1843431f709345e9b51f51be76a6c3aed2e179