Treat CID on cmdline as signed
We pass a CID value of -1U (VMADDR_CID_ANY) from composd to odrefresh
to pvm_exec as a sentinel value to indicate use of
composd. Unfortunately odrefresh expects it to be signed while composd
and pvm_exec expect it to be unsigned, leading to test breakge.

This CL moves everything to signed. (Technically unsigned is more
correct, but -1 looks more elegant than 4294967295.

Fixes: 202696349
Test: atest ComposTestCase
Test: adb shell apex/com.android.compos/bin/composd_cmd
Change-Id: I9f8cbe85ebf6755a158c20de95c62e3d44591db0