Use cid=0 to mean not specified
We're using -1 (VMADDR_CID_ANY) as a special value to mean "use
composd rather than directly connecting to a specific CID". So we
can't also use it to mean "no CID was specified, run dex2oat
directly".

The value 0 is never valid as a CID, since it's reserved as
VMADDR_CID_HYPERVISOR.

Eventually the CID parameter can just go away, and we'll have a
boolean for use composd in a VM / normal no VM.

Bug: 202696349
Test: atest ComposTestCase
Test: adb shell apex/com.android.compos/bin/composd_cmd
Test: Manually verify compilation runs in the VM.
Change-Id: I11a4814481576a62294701ed2bb2f75808a4ef71