[scudo] Skip AllocAfterFork test on machines with low max_map_count
Reducing the number of iterations in that test with D111342 helped,
but the failure still occured flakily when the test is ran as part
of a large test suite.

Reducing further the number of iterations might not be good enough,
so we will skip the test if the `max_map_count` variable can be
read, and if lower than a given threshold.

Differential Revision: https://reviews.llvm.org/D111465

GitOrigin-RevId: 56a9effc427b10203dfb7f691a12924a82ba5cd5
Change-Id: I36c7abf6c2f96b7bcc32f22014befda0afbc66d2