Add a flag to control linter behaviour.
'cros lint' lints entire files. However, many Chrome OS seccomp
policies will fail the linter. This will generate a ton of noise for
Chrome OS developers.

Because the linter only prints output and blocks submission when the
underlying tool returns nonzero, add a flag to control this. That way
we can control the behaviour of the linter from cros_lint.py, and not
have to uprev Minijail to change this behaviour.

It was a bug that the linter was returning zero unconditionally. That
probably masked the fact that turning the linter on would fail on many
policies (because it's checking full files and not just diffs.)

We will likely have to fix some of the existing policies before turning
the flag on.

Bug: 202190205
Test: $ tools/seccomp_policy_lint.py seccomp.policy --failures-return-nonzero; echo $?
Test: ...
Test: 1
Change-Id: Ia82e0a13e547a61852c62283ca79df9097f6cde7