Autoupdate of 3 submodules
*build/bazel updated to 98f8a587ad7b69b95a105449c01464b3f3dcf740
Convert cpp_std to use toolchain features.

Also removed usage of kwargs to create the common attrs for the cpp
intermediate targets, because it was silently overwriting the features
with an empty features list from kwargs in the dict construction.

Fixes: 202518741
Test: CI
Change-Id: I3acdcffaeab340ac19bdffc3551ac677a6f58546

*build/soong updated to e466cc7ea65406295613e4cd21033e1e2e75c435
Merge "Convert cpp_std to be an attribute instead of a copt."
*prebuilts/clang/host/linux-x86 updated to 52b9a88df9dfbc8d946f835ddc83269081b546ec
Merge "Convert cpp_std to features."