Add minimal support for maxTargetSdk=S.
This allows developers to put @UnsupportedAppUsage(maxTargetSdk=S) in
their code, but does not have the correct behaviour yet. Any APIs
annotated like that will remain on the unsupported list (i.e. not
actually restricted).

The remaining work is tracked in b/203045121, to create the new list
and enforce access appropriately.

Bug: 169748350
Bug: 202849738
Test: TH
Change-Id: I6cbc554299d53f6d243c90e7915f0624117cf1a2