Get the updated link state from netlink event instead of Netd.
Get the updated link state from ifi_flags of ifinfomsg header, and
sync up with IpClient on the link state change.

Also do the initialization of mDependencies in IpClientLinkObserver
prior to mNetlinkMonitor. The purpose is to prevent the NPE when
calling isNetlinkEventParsingEnabled in MyNetlinkMonitor ctor. This
change is safe, just adjusting the initialization order of private
members, it doesn't bring any side-effect.

Bug: 163492391
Test: atest NetworkStackTests NetworkStackIntegrationTests
Change-Id: I6bf5929fefd00526c439178defba2ac2597682d0