Revert "Revert "Create a generated delegator class""
This reverts commit 275fca8b337952a5498762a233a775c814722038.

The initial commit updated some golden_output files with a repo that was
out-of-date. It was missing some recent system/tools/aidl changes that
also updated the golden files. Presubmits did not catch this difference.

This change is rebased with the latest changes and the golden_output
files reflect that.

Test: atest aidl_integration_test aidl_unittests binderRpcTests
Bug: 198785815
Change-Id: Id16799305ae8bf5daae583bf1f5ecdea11591261