Break the SharedLog → LocalLog dependency.
LocalLog is an internal class to the framework that should
not be depended upon by libs. Today it's used as a filegroup.
Instead, copy the relevant code, and thin down everything
that wasn't actually needed, making the addition lean.

Unfortunately this doesn't make it possible to remove
LocalLog from any target, because all actual clients of
SharedLog *also* have a separate dependency to LocalLog.
This includes NetworkStack, CellBroadcastReceiver,
Scheduling, Wifi and maybe others.

Still, this is a good step and it makes it possible to move
client of SharedLog to a better place.

Test: SharedLogTest
Change-Id: I7b002ad0ee20b7d4efca11402bbc5c3f48f75437