[DO NOT MERGE] CTS 10: Work around blocklist class name change.
Part of a set of changes to Android 10 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

This change is a backport of the equivalent CTS 11 change
in https://r.android.com/1833498

In the Android 12 based Conscrypt module, the certificate blocklist
class was renamed, but the old class name is used in the Android 10 CTS
suite.

As this only affects a single test and as this test is exercised by
Conscrypt MTS before the module can be pushed, simply skip the test if
the expected class is unavailable.

On an Android 10 device with an earlier Conscrypt module, the test
will run as before.

Required upgrading the test to Junit 4 to allow assumptions to work
correctly.

Bug: 197500794
Bug: 195615983
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt
module

Change-Id: I3962c5b389538d20206a91cf360a19eff47d1e58