base: Use timerfd in Watchdog and serialize crash logs
This CL changes the way the base::Watchdog works.
Before: each task would create a new posix timer which,
on expiration, would deliver a SIGABRT.

After: the watchdog thread keeps one timerfd for all tasks
(keeping it armed to the closest task) and monitors
expired timers in the watchdog thread loop.

This is to have a chance to call
base::MaybeSerializeLastLogsForCrashReporting() before
dispatching a SIGABRT, so that the last logs (and soon
the crash keys) are available.

Bug: 202283508
Test: perfetto_unittests --gtest_filter=WatchdogTest*
Change-Id: I38db82a13d882c00be96273ef2f334d82f9eadcb