Remove sqlite-devel from Linux Dockerfile
The NDK build of Python3 probably doesn't need sqlite.

Windows explicitly omits the _sqlite3 Python3 module because it's not
needed. The macOS build currently also lacks the _sqlite3 module, but
for incidental reasons. (i.e. The macOS target is 10.9, and Python
tries to use APIs that are only in macOS 10.10+, such as
sqlite3_close_v2.)

Bug: none
Test: external/python/cpython3/kokoro_build.sh
Change-Id: I17ac2d00cb9e09185754db1fbb3a03bfde608e04