Bump zic version to 2021d.
This commit updates the IANA tzcode binary used on host to generate the
tzdata (TZif archive) file used on Android by bionic and the
java.util.TimeZone implementation. This update does not alter the code
used on the device at runtime, though changes to the content of the file
can affect device behavior.

Note: There are third party time zone libraries that utilize the tzdata
file, but these are not supported by the Android platform team.

This commit should not be backported to earlier Android releases unless
later bug fixes rely on it, as it may affect behavior / app
compatibility.

Notes for Android time zone maintainers:
Like tzdb updates, the binary files in this commit should not be patched
to other release branches: they are not guaranteed to work.

According to dump-tzdata.py there were 2 types of changes:
  1) Some time zones had extra no-effect time transition entry. In case
  of Europe/London there was 1996-01-01T00:00:00 transition, which
  corresponds to 'GB-Eire' -> 'EU' switch in zone rules [1]

  2) isutcnt/isstdcnt are set to 0 if values in standard/wall / UT/local
  indicators were all zeros. Specification allows that [2].
  I've checked few time zone files manually to verify that.
  ZoneInfoData does not read them, so this change should not affect
  libcore.

Also, by default slim TZif files are generated, so '-b fat' option was
passed to zic.

Bug: 203414801
Test: CtsLibcoreTestCases
Test: CtsLibcoreOjTestCases
Test: CtsIcuTestCases
Test: CtsBionicTestCases
Test: CtsTextTestCases

[1]
https://github.com/eggert/tz/blob/951b4f301bc01703692d2d0017a17825dc33cb2d/europe#L506

[2] https://datatracker.ietf.org/doc/html/rfc8536#section-3.2

Change-Id: I25c041fa8cdfab0b23d24ac73f109a15cdf0b391