Autoupdate of 10 submodules
*device/generic/goldfish updated to 1361dec1fdc71c3bacff69fcc8bf96b1884e25d9
Merge "Revert "Revert "Remove the sysfs_block_type SELinux attribute"""
*device/google/bonito-sepolicy updated to 37ae749c20a8027e2c6042044180eb1bdbc5c4f2
Revert "Revert "Remove the bdev_type and sysfs_block_type SELinux attributes""

This reverts commit f8e7785812eb970baa335b81e9254fa791ea9f1f.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Ia2761e819bd152666bad922d22e36505beba9a28

*device/google/coral-sepolicy updated to 91e85bd0fb5a6a62f377035dcc2993d1179eb350
Revert "Revert "Remove the bdev_type and sysfs_block_type SELinux attributes""

This reverts commit c68582718479f09b9cc0ebf60cb1e129ff1e679a.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: I190c158d5f31c969595e34160185bc929d6e8d69

*device/google/crosshatch-sepolicy updated to 7dc695097be0e0af6497eae80efa474d21652fc0
Revert "Revert "Remove the bdev_type and sysfs_block_type SELinux attributes""

This reverts commit ceb89d2986ff2f6b6a1f055641fbe9ae6730a4c3.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Ia8fc800c6180c3264e602b1ae60ca857fd003d9c

*device/google/cuttlefish updated to 443c8f214cd8e1fca250e2921767a82a8ae1727a
Merge "Revert "Revert^2 "Add the 'bdev_type' attribute to all block devices"""
*device/google/redbull-sepolicy updated to 98fd7e8458ff8ffe16aef420cbb09893294ff5e0
Revert "Revert "Remove the bdev_type and sysfs_block_type SELinux attributes""

This reverts commit 45bc87f69bf766e1e0e100346ec4107f00b94b9d.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Iba682eaac53d522da06fa3f3f9e86dd384e36f42

*device/google/sunfish-sepolicy updated to ef2f3aec105762a5f893a7e8cbb218d88bf26138
Revert "Revert "Remove the bdev_type and sysfs_block_type SELinux attributes""

This reverts commit 2ce2569b3507b20f6f733fcc6dcf79e72c66682d.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: I640160b0c310e67fed6cf1374bf0fdfcbfdd5e1e

*device/linaro/dragonboard updated to 307752944c952946a3c0e91c776d3c33de873be6
Revert "Revert^2 "Add the 'bdev_type' attribute to all block devices""

This reverts commit 03b647d2055957ad9c92a3e9c7c60a93a1807199.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Ia2629994551d4f9ca12acc109104472450db98dc

*device/linaro/poplar updated to f6af322852f7cf621e8c447ad45072d225ed0b8d
Revert "Revert^2 "Add the 'bdev_type' attribute to all block devices""

This reverts commit c4af45caa3cdc91c236a59ae2058968732734c8c.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Ieb3346ecc604f4365dcace125072c1927c7a647c

*device/mediatek/wembley-sepolicy updated to e46e4134e5041d923142d2b4e5c2949c9cf821a1
Revert "Revert "Revert "Add the 'bdev_type' attribute to all block devic...""

This reverts commit 5503dda0d76103d9ee9dafaac408518c3bebffe4.

Reason for revert: Restore this patch since it was not necessary to revert this patch.
Bug: 202520796
Change-Id: Ie3a76b4beb3b58ab849b52c796b962981d3e12fc