Change Version::CURRENT_RUNTIME to getCurrentRuntimeVersion()
Prior to this change, Version::CURRENT_RUNTIME represented a version of
a feature that is not part of any HAL specification (e.g., Control Flow
operations with operands of unknown size). Because of this, the runtime
version is always the latest available version + 1.

Version::CURRENT_RUNTIME does not work as a constant when the latest
available version is not a constant. For example, the latest available
version will be variable with the addition of Experimental Feature Flags
to the NNAPI. This change creates a helper function
"getCurrentRuntimeVersion" to return the variable current runtime
version in preparation for Experimental Feature Flags.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Change-Id: I701d9ca248a815b43cc24b14b24358ba429b0dd4