BTM resolve race condition in ble local resolving list
The BLE advertiser address resolving list is stored on the controller,
a local copy is kept on the host. The updates to the local copy did not
always match the remote list.

In this change the update of the local list for the add operation
is moved to the callback in order to ensure the order of operations at
the controller is respected and that the operation was successful.

Bug: 119574882
Bug: 199827901
Test: gd/cert/run
Tag: #stability
Change-Id: Ib215590e4a292b8d04fcb1a46c13b464d517c73c
(cherry picked from commit 42059daf9be75ac3daeda5955b9759584a439ae6)
(cherry picked from commit fb070a9fd45bfdc6e452c9735649526d4ba6b65e)