Allow for unsigned byte constants (u8)
For int32 constants, we can use something like 0xffffffff in order to
express a 32-bit value. However, byte literals are automatically
converted to 32-bit values. This is consistent with how integral
promotion works in C. However, this can be annoying when trying to
create 8-bit constants. So, the integer literal suffix 'u8' is added in
order to represent byte literals.

Fixes: 203492856
Test: aidl_integration_test, aidl_unittests
Change-Id: Idc9c842306d6410cc979e21b6ad5d3423a87354d