Update Android ICU tzdata from 2021a1 to 2021a3.
This change is associated with tzdb 2021a rev. 3 changes in
system/timezone.

Please keep in mind that ICU's reported version will be 2021a, not
2021a3 even though technically it is 2021a1.

zoneinfo64.txt and binary files are generated as part of running
system/timezone/update-tzdata.py

This CLs contains three changes from upstream:
  1) https://github.com/unicode-org/icu/pull/1912
  2) https://github.com/unicode-org/icu/pull/1916
  3) https://github.com/unicode-org/icu/pull/1920

TimeZoneTest was changed: America/Panama is not canonical
representation of America/Panama as we dropped that from backward file.

Bug: 204052899
Test: see system/timezone CL

Change-Id: I9442dfc4e08cd9dc55b5afe5ed5e4d9336dadfa5