Autoupdate of 2 submodules
*device/mediatek/wembley-sepolicy updated to 138e227156e55ce0e1f91af6d4967d3a012be8fa
Remove logic that duplicates system policy

With https://android-review.googlesource.com/c/platform/system/sepolicy/+/1874738
we are fixing init's /proc/cpu/alignment write access for the
entire system.  So we remove this device-specific version of it
to allow the policy to continue to build.

Bug: 145371497
Test: Local build and boot
Change-Id: Ic67ba5f9428b7a3da704a8b40474212f02ab6667

*system/sepolicy updated to 54bd8438b1ba8020a66720c1f8a14f5b0b2c4770
Merge "Allow init to write to /proc/cpu/alignment"