suspend: Avoid releasing and retaking lock
Some error conditions in the autosuspend loop occur while
the autosuspendLock is held. These errors cause the loop
to move to the next iteration, effectively causing the
autosuspendLock to be release (drops out of scope) and
immediately recreated and locked at the beginning of the
next iteration of the loop. Avoid this unecessary
unlocking-relocking by moving the lock outside the loop

Change-Id: I61c7675e8041ddef6225d81f4a08350a8b8ffa34
Test: SystemSuspendV1_0UnitTest
Bug: N/A
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Suggested-by: Steven Moreland <smoreland@google.com>