Reject unstructured parcelables in NDK/Rust
There's no way to get their definitions (or locations) in NDK/Rust. In
C++, `cpp_header` tells its header and in Java it's assumed to be found
in classpath.

If we need to support them in NDK/Rust, we can add a new keyword
`ndk_header` for NDK header and `rust_crate` for Rust crate for them.

Fixes: 204817775
Test: aidl_unittest
Change-Id: I10f26502de8a8a087adede7fa6a16757a25801a1