Improve error reporting of waitForPerfettoIdle
Integration tests on the CI are flaking because waitForPerfettoIdle
fails, and it does not show a particularly helpful error message when
that happens. This CL improves its reporting, adding human-readable list
of reasons why Perfetto UI is not considered to be idle. This hopefully
should aid in debugging integration test flakes on the CI.

Bug: 202738850
Change-Id: I180c7cf73430387464c15b65447db9e95e6fa768