Image verifier: Added flexibility to use curl instead of wget.
Bug: 205021723

On some systems, `wget` is not pre-bundled, but `curl` is. This
change checks if either of the tools is installed and used whichever
that's available to it. If both are available, `wget` is used.

Also unpegged `fbpacktool` from being located strictly at where
`avbtool` is. As long as it is in $PATH, we will proceed. Else,
we will fail early and not wait until download and unzipping
is complete.

Test:
`python3 pixel_factory_image_verify.py https://dl.google.com/dl/android/aosp/raven-sd1a.210817.036.a8-factory-7352ca24.zip`
Successfully fetched, unpacked, verified, and created payload.txt.

Change-Id: Ie1a3216e044de25ff3318a4712f1092de47a106f