Store the control path in a test-specific dir
Relocate the generated SSH config to a test-specific directory such that
permission issues are resolved.

When running ACTS on the same machine with several users, keeping the
generated SSH config in a common place causes permissions issues where
one user is not able to read if another user created the file.

Fixes https://fxbug.dev/87942.

Bug: None
Test: act.py [...] -tc Sl4fSanityTest
Change-Id: I994b3af1f810cb9680f16de8d3de53c277728d5b