Verify bootconfig partition when locked
If CONFIG_AVB_IS_UNLOCKED is not set, the bootconfig partition is
verified using AVB. This is to prevent attackers from being able to pass
arbitrary bootconfigs which can be used to change the behavior of
Android.

The bootconfig partition has to be signed with `avbtool add_hash_footer`
and the partition has to registered as a chained partition from the main
vbmeta partition. This allows us to add more verifiable partitions to
the disk.

Bug: 203031847
Test: run microdroid
Change-Id: Ic68f620b05257098158766e9b49fac7f8e5274ea