base: Small fix UnixSocket::SendStr testing method
A small change to the Send(const std::string&) method of
UnixSocket, which is used only in tests. The current method
appends a null-terminator to the string sent. This makes it
impossible to use with cases (upcoming in future CLs) where
we want to just send a larger buffer in chunks, as it injects
unwanted \0 in the middle.

This method seems used only in unix_socket_unittest.cc. Even
there, most usages are fine because ReceiveString already
null-terminates the received string.
In order to avoid accidental behavioral changes in code outside
of the repo, I am changing the name of the method, so if there
is any subtle dependency on the old behavior it will break
rather than silently omit terminators.

Test: perfetto_unittests
Bug: 205274609
Change-Id: Ia17ad2c10ad8979b5dc241b0387e41b7189ec519