Autoupdate of 2 submodules
*system/extras updated to f1ce5ed2a95c864ba06cade6556daa9f0f323bb9
Merge "Update for libunwindstack changes."
*system/unwinding updated to 734cc2f1d45d5e0269654cb5b8fdecfc93dd2ad3
Remove the next_real_map and prev_real_map.

It's a pain to keep track of these pointers when doing a reparse.
Rather than keeping these in sync, remove them completely and add
a next_map. The next_real_map and prev_real_map variables are only
needed in very specific circumstances that are very unlikely to
occur. In those cases, derive the values when they are needed.

Also, make the Get{Next,Prev}RealMap functions verify that the
name is the same so that those checks can be removed from the
places that use those values.

Add new unit tests to verify the new behavior, and verify the
pointers are set properly.

Test: All unit tests pass.
Change-Id: I8fd784df71d7b39890c763755bec5904c0bed8a3