[llcpp][goldfish] Migrate goldfish to LLCPP's inline table setters
FIDL is undergoing a migration to a new wire format that inlines small
(<= 4 byte) table fields in FIDL envelopes. Because LLCPP is a low level
binding, the API is also updating to directly inline table fields
without use of an arena. This CL updates deprecated usages of arena
setters to facilitate this migration.

Bug: fxbug.dev/79578
Change-Id: Iec73825fc990fa63073a16e1cc9a08019ff0ffc1